requirements.txt
pinecone-client==2.2.4
openai==1.3.0
python-dotenv==1.0.0
rhinomcp==0.1.0 